+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.1.4
+
+- Parcel CSS -> Lightning CSS
+- Sort keys for CSSModuleExports
+
+## 0.1.3
+
+Bump config-loader
+
 ## 0.1.2
 
 Fix issue with inset-x/y